+++ assets/js/blocks/cart-checkout/cart/inner-blocks/cart-order-summary-block/frontend.tsx
@@ -4,6 +4,9 @@
 import { __ } from '@wordpress/i18n';
 import Title from '@woocommerce/base-components/title';
 import { withFilteredAttributes } from '@woocommerce/shared-hocs';
+import { TotalsFooterItem } from '@woocommerce/base-components/cart-checkout';
+import { getCurrencyFromPriceResponse } from '@woocommerce/price-format';
+import { useStoreCart } from '@woocommerce/base-context/hooks';
 
 /**
  * Internal dependencies
@@ -18,12 +21,19 @@ const FrontendBlock = ( {
 	children: JSX.Element | JSX.Element[];
 	className: string;
 } ): JSX.Element | null => {
+	const { cartTotals } = useStoreCart();
+	const totalsCurrency = getCurrencyFromPriceResponse( cartTotals );
+
 	return (
 		<div className={ className }>
 			<Title headingLevel="2" className="wc-block-cart__totals-title">
 				{ __( 'Cart totals', 'woo-gutenberg-products-block' ) }
 			</Title>
 			{ children }
+			<TotalsFooterItem
+				currency={ totalsCurrency }
+				values={ cartTotals }
+			/>
 			<OrderMetaSlotFill />
 		</div>
 	);
