--- assets/js/blocks/cart-checkout/cart/inner-blocks/cart-order-summary-block/total/block.tsx
@@ -1,21 +1,0 @@
-/**
- * External dependencies
- */
-import { TotalsFooterItem } from '@woocommerce/base-components/cart-checkout';
-import { getCurrencyFromPriceResponse } from '@woocommerce/price-format';
-import { useStoreCart } from '@woocommerce/base-context/hooks';
-
-const Block = ( { className }: { className: string } ): JSX.Element => {
-	const { cartTotals } = useStoreCart();
-	const totalsCurrency = getCurrencyFromPriceResponse( cartTotals );
-
-	return (
-		<TotalsFooterItem
-			className={ className }
-			currency={ totalsCurrency }
-			values={ cartTotals }
-		/>
-	);
-};
-
-export default Block;
