--- assets/js/blocks/cart-checkout/cart/inner-blocks/cart-order-summary-block/total/edit.tsx
@@ -1,30 +1,0 @@
-/**
- * External dependencies
- */
-import { useBlockProps } from '@wordpress/block-editor';
-
-/**
- * Internal dependencies
- */
-import Block from './block';
-
-export const Edit = ( {
-	attributes,
-}: {
-	attributes: {
-		className: string;
-	};
-	setAttributes: ( attributes: Record< string, unknown > ) => void;
-} ): JSX.Element => {
-	const { className } = attributes;
-	const blockProps = useBlockProps();
-	return (
-		<div { ...blockProps }>
-			<Block className={ className } />
-		</div>
-	);
-};
-
-export const Save = (): JSX.Element => {
-	return <div { ...useBlockProps.save() } />;
-};
