--- assets/js/blocks/cart-checkout/cart/inner-blocks/cart-order-summary-block/total/frontend.tsx
@@ -1,6 +1,0 @@
-/**
- * Internal dependencies
- */
-import Block from './block';
-
-export default Block;
