--- assets/js/blocks/cart-checkout/cart/inner-blocks/cart-order-summary-block/total/index.tsx
@@ -1,25 +1,0 @@
-/**
- * External dependencies
- */
-import { totals } from '@woocommerce/icons';
-import { Icon } from '@wordpress/icons';
-import { registerFeaturePluginBlockType } from '@woocommerce/block-settings';
-
-/**
- * Internal dependencies
- */
-import { Edit, Save } from './edit';
-import metadata from './block.json';
-
-registerFeaturePluginBlockType( metadata, {
-	icon: {
-		src: (
-			<Icon
-				icon={ totals }
-				className="wc-block-editor-components-block-icon"
-			/>
-		),
-	},
-	edit: Edit,
-	save: Save,
-} );
